services/auth/pyproject.toml
[project]
name = "dsdc"
version = "0.1.0"
description = "DataScientest Document Classification Project"
requires-python = ">=3.12"
readme = "README.md"
license = "MIT"
authors = [
  { name = "AL"},
  { name = "BH"},
]

dependencies = [
    "argon2-cffi>=25.1.0",
    "fastapi>=0.116.1",
    "passlib>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
#    "requests>=2.32.5",
    "uvicorn[standard]>=0.35.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["dsdc*"]

[dependency-groups]
dev = [
    "ipython>=9.5.0",
    "pytest>=8.4.2",
]
